requirements.txt
google-generativeai
python-dotenv
pyyaml
tiktoken
